specify node dependency in nvim_wrapper_config rather than chosen_tools
if I ever work on a project with a node dependency, the prior setup
could cause conflict between the node from the flake and the node the
project needs. This resolves that by bundling an nvim-node binary as
part of installing nvim which is automatically set as nvim's default
node to use.